// scripts/importFromOldDB/libs/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type Company = {
  __typename: "Company",
  id?: string,
  name?: string,
  startMonth?: number,
  shortName?: string | null,
  url?: string | null,
  createdAt?: string,
  updatedAt?: string,
};

export type UpdateCompanyInput = {
  id: string,
  name?: string | null,
  startMonth?: number | null,
  shortName?: string | null,
  url?: string | null,
};

export type DeleteCompanyInput = {
  id?: string | null,
};

export type DeleteEmployeeInput = {
  username: string,
};

export type ModelEmployeeConditionInput = {
  companyID?: ModelIDInput | null,
  no?: ModelStringInput | null,
  groupID?: ModelIDInput | null,
  superiorUsername?: ModelStringInput | null,
  firstName?: ModelStringInput | null,
  lastName?: ModelStringInput | null,
  grade?: ModelStringInput | null,
  email?: ModelStringInput | null,
  sub?: ModelStringInput | null,
  manager?: ModelEmployeeTypeInput | null,
  isCompanyAdmin?: ModelBooleanInput | null,
  isDeleted?: ModelBooleanTypeInput | null,
  and?: Array< ModelEmployeeConditionInput | null > | null,
  or?: Array< ModelEmployeeConditionInput | null > | null,
  not?: ModelEmployeeConditionInput | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelEmployeeTypeInput = {
  eq?: EmployeeType | null,
  ne?: EmployeeType | null,
};

export enum EmployeeType {
  OTHER = "OTHER",
  NORMAL = "NORMAL",
  MANAGER = "MANAGER",
  SUPER_MANAGER = "SUPER_MANAGER",
}


export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelBooleanTypeInput = {
  eq?: BooleanType | null,
  ne?: BooleanType | null,
};

export enum BooleanType {
  TRUE = "TRUE",
  FALSE = "FALSE",
}


export type Employee = {
  __typename: "Employee",
  username?: string,
  companyID?: string,
  no?: string,
  groupID?: string,
  superiorUsername?: string,
  firstName?: string,
  lastName?: string,
  grade?: string,
  email?: string,
  sub?: string | null,
  manager?: EmployeeType,
  isCompanyAdmin?: boolean | null,
  isDeleted?: BooleanType,
  createdAt?: string,
  updatedAt?: string,
  company?: Company,
  superior?: Employee,
  group?: Group,
};

export type Group = {
  __typename: "Group",
  id?: string,
  companyID?: string,
  no?: string,
  name?: string,
  createdAt?: string,
  updatedAt?: string,
};
// ...
export type CreateEmployeeInput = {
  username: string,
  companyID: string,
  no: string,
  groupID: string,
  superiorUsername: string,
  firstName: string,
  lastName: string,
  grade: string,
  email: string,
  sub?: string | null,
  manager: EmployeeType,
  isCompanyAdmin?: boolean | null,
  isDeleted: BooleanType,
};

export type UpdateEmployeeInput = {
  username: string,
  companyID?: string | null,
  no?: string | null,
  groupID?: string | null,
  superiorUsername?: string | null,
  firstName?: string | null,
  lastName?: string | null,
  grade?: string | null,
  email?: string | null,
  sub?: string | null,
  manager?: EmployeeType | null,
  isCompanyAdmin?: boolean | null,
  isDeleted?: BooleanType | null,
};
// ...
export type ModelStringKeyConditionInput = {
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelEmployeeFilterInput = {
  username?: ModelStringInput | null,
  companyID?: ModelIDInput | null,
  no?: ModelStringInput | null,
  groupID?: ModelIDInput | null,
  superiorUsername?: ModelStringInput | null,
  firstName?: ModelStringInput | null,
  lastName?: ModelStringInput | null,
  grade?: ModelStringInput | null,
  email?: ModelStringInput | null,
  sub?: ModelStringInput | null,
  manager?: ModelEmployeeTypeInput | null,
  isCompanyAdmin?: ModelBooleanInput | null,
  isDeleted?: ModelBooleanTypeInput | null,
  and?: Array< ModelEmployeeFilterInput | null > | null,
  or?: Array< ModelEmployeeFilterInput | null > | null,
  not?: ModelEmployeeFilterInput | null,
};

export type ModelEmployeeConnection = {
  __typename: "ModelEmployeeConnection",
  items?:  Array<Employee | null > | null,
  nextToken?: string | null,
};

export type ModelEmployeeEmployeeByCompanyManagerCompositeKeyConditionInput = {
  eq?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null,
  le?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null,
  lt?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null,
  ge?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null,
  gt?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null,
  between?: Array< ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null > | null,
  beginsWith?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput | null,
};

export type ModelEmployeeEmployeeByCompanyManagerCompositeKeyInput = {
  manager?: EmployeeType | null,
  isDeleted?: BooleanType | null,
};
// ...
export type DeleteEmployeeMutationVariables = {
  input?: DeleteEmployeeInput,
  condition?: ModelEmployeeConditionInput | null,
};
// ...
export type CreateEmployeeMutationVariables = {
  input?: CreateEmployeeInput,
  condition?: ModelEmployeeConditionInput | null,
};
// ...
export type UpdateEmployeeMutationVariables = {
  input?: UpdateEmployeeInput,
  condition?: ModelEmployeeConditionInput | null,
};
// ...
export type ListEmployeesCompanyQueryVariables = {
  companyID?: string | null,
  no?: ModelStringKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelEmployeeFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};
// ...
export type ListEmployeesManagerQueryVariables = {
  companyID?: string | null,
  managerIsDeleted?: ModelEmployeeEmployeeByCompanyManagerCompositeKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelEmployeeFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};
// ...
export type GetEmployeeQueryVariables = {
  username?: string,
};